Add files via upload
This is the fully styled version of the app. A hero image and background image were added, a grey color scheme was chosen, and the CSS has been set to respond to screen sizes up to 1440 pixel width.

This still needs to be tested in github pages before it's completely ready.